Auth0 Integration (#32)
* Auth0 Integration

This PR for the Integrate Auth0 logs into Wazuh by polling the Auth0 Management API. Logs such as successful logins, failed logins, password resets, and suspicious activity will be collected from Auth0 and forwarded to Wazuh for monitoring, alerting, and correlation. Developed and tested on Wazuh server v4.12.

* Update README.md

* Updated the folder name

* Update README.md

* Update README.md

* Add files via upload

* Screenshots and Scripts update

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* OpenVAS Integration

This PR for the OpenVAS (part of the Greenbone Vulnerability Management suite) is an open-source vulnerability scanner that identifies security risks across systems and applications. Wazuh, on the other hand, is a Security Information and Event Management (SIEM) and XDR platform that centralises log collection, threat detection, and alerting.
Developed and tested on Wazuh server v4.12.

* Update README.md

* screenshots update

* Update README.md

* Created openvas_extract shell script

* updating the existing auth0 repository

Adding the README.md file with scripts and Screenshots

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Create openvas_rules.xml

Rules for open ports and CVEs

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Delete integrations/auth0_Integration directory

Moved all content to existing auth0 repository

* Update README.md

update the screenshots path

* Update README.md

* Remove Openvas Integration

* Update README.md

* Update README.md

* Update README.md

* Update README.md